moved more core logic into the version api - update core documentation to be more simplified - updated version api documentation to be more in-depth
