+++ pkg/core/predict.go
@@ -10,15 +10,10 @@ type PredictVersionOptions struct {
 	SemverMode     string
 }
 
-// PredictVersion predicts a version based on the latest annotated git tag and a map of semver levels
-// matched to specific strings.
-// Returns the predicted version or an error if anything went wrong with the increment.
+// PredictVersion predicts the next version.
+// Returns the predicted version or an error if anything went wrong with the prediction.
 func PredictVersion(options *PredictVersionOptions) (prediction string, err error) {
 	var versionAPI = versions.NewAPI()
 	var version = versionAPI.GetVersionOrDefault(options.DefaultVersion)
-	return versionAPI.PredictVersion(
-		version,
-		options.SemverMatchMap,
-		options.SemverMode,
-	)
+	return versionAPI.PredictVersion(version, options.SemverMatchMap, options.SemverMode)
 }
